# swagger.yaml
openapi: 3.0.0
paths:
...
  /api/v1/users:
    get:
      operationId: UsersController_find
      summary: Get users paginated list
      parameters:
        - name: filters
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                filters[id][$eq]: 1
          schema:
            type: object
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            type: integer
        - name: pageSize
          required: false
          in: query
          schema:
            minimum: 1
            type: integer
        - name: sorts
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                sorts[id]: asc
          schema:
            type: object
        - name: includes[]
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - articles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
      tags: &ref_6
        - Users
      security:
        - bearer: []
    post:
      operationId: UsersController_create
      summary: Create new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
      tags: *ref_6
      security:
        - bearer: []
  /api/v1/users/all:
    get:
      operationId: UsersController_findAll
      summary: Get all users
      parameters:
        - name: filters
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                filters[id][$eq]: 1
          schema:
            type: object
        - name: sorts
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                sorts[id]: asc
          schema:
            type: object
        - name: includes[]
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - articles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
      tags: *ref_6
      security:
        - bearer: []
  /api/v1/users/{id}:
    get:
      operationId: UsersController_findOne
      summary: Get specific user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
        - name: includes[]
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - articles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_6
      security:
        - bearer: []
    put:
      operationId: UsersController_update
      summary: Update specific user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_6
      security:
        - bearer: []
    delete:
      operationId: UsersController_remove
      summary: Delete specific user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_6
      security:
        - bearer: []
info:
  title: Nest Articles App
  description: RESTful API with NestJS and TypeORM
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3333
    description: Dev Server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        rowCount:
          type: integer
          minimum: 0
        pageCount:
          type: integer
          minimum: 0
      required:
        - page
        - pageSize
        - rowCount
        - pageCount
...
    User:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
        name:
          type: string
          maxLength: 100
        username:
          type: string
          maxLength: 100
        email:
          type: string
          maxLength: 100
          format: email
      required:
        - id
        - created_at
        - updated_at
        - name
        - username
        - email
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        username:
          type: string
          maxLength: 100
        password:
          type: string
          minLength: 6
          maxLength: 50
        email:
          type: string
          maxLength: 100
          format: email
      required:
        - name
        - username
        - password
        - email
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        username:
          type: string
          maxLength: 100
        password:
          type: string
          minLength: 6
          maxLength: 50
        email:
          type: string
          maxLength: 100
          format: email
      required:
        - name
        - username
        - email